refactor(ResizeDetector): Move the resize detector to the top (#382)
